requirements.txt
MetaTrader5
pandas
numpy
matplotlib
pyyaml
requests
